refactor(go): colocate authCapture escrow ABI and nonce helpers
Move ComputeAuthCaptureNonce, PaymentInfoTypehash, and AuthCaptureEscrow
contract ABIs from package evm into authCapture/core for parity with
batch-settlement packaging. HashReceiveWithAuthorization stays in eip712.go.
