+++ go/mechanisms/evm/constants.go
@@ -774,105 +774,4 @@ var ReceiveWithAuthorizationBytesABI = []byte(`[
 	}
 ]`)
 
-// AuthCaptureEscrow ABI for authorize function
-var EscrowAuthorizeABI = []byte(`[
-	{
-		"inputs": [
-			{
-				"name": "paymentInfo",
-				"type": "tuple",
-				"components": [
-					{"name": "operator", "type": "address"},
-					{"name": "payer", "type": "address"},
-					{"name": "receiver", "type": "address"},
-					{"name": "token", "type": "address"},
-					{"name": "maxAmount", "type": "uint120"},
-					{"name": "preApprovalExpiry", "type": "uint48"},
-					{"name": "authorizationExpiry", "type": "uint48"},
-					{"name": "refundExpiry", "type": "uint48"},
-					{"name": "minFeeBps", "type": "uint16"},
-					{"name": "maxFeeBps", "type": "uint16"},
-					{"name": "feeReceiver", "type": "address"},
-					{"name": "salt", "type": "uint256"}
-				]
-			},
-			{"name": "amount", "type": "uint256"},
-			{"name": "tokenCollector", "type": "address"},
-			{"name": "collectorData", "type": "bytes"}
-		],
-		"name": "authorize",
-		"outputs": [],
-		"stateMutability": "nonpayable",
-		"type": "function"
-	}
-]`)
-
-// AuthCaptureEscrow ABI for charge function
-var EscrowChargeABI = []byte(`[
-	{
-		"inputs": [
-			{
-				"name": "paymentInfo",
-				"type": "tuple",
-				"components": [
-					{"name": "operator", "type": "address"},
-					{"name": "payer", "type": "address"},
-					{"name": "receiver", "type": "address"},
-					{"name": "token", "type": "address"},
-					{"name": "maxAmount", "type": "uint120"},
-					{"name": "preApprovalExpiry", "type": "uint48"},
-					{"name": "authorizationExpiry", "type": "uint48"},
-					{"name": "refundExpiry", "type": "uint48"},
-					{"name": "minFeeBps", "type": "uint16"},
-					{"name": "maxFeeBps", "type": "uint16"},
-					{"name": "feeReceiver", "type": "address"},
-					{"name": "salt", "type": "uint256"}
-				]
-			},
-			{"name": "amount", "type": "uint256"},
-			{"name": "tokenCollector", "type": "address"},
-			{"name": "collectorData", "type": "bytes"},
-			{"name": "feeBps", "type": "uint16"},
-			{"name": "feeReceiver", "type": "address"}
-		],
-		"name": "charge",
-		"outputs": [],
-		"stateMutability": "nonpayable",
-		"type": "function"
-	}
-]`)
-
-// AuthCaptureEscrow ABI for getHash function
-var EscrowGetHashABI = []byte(`[
-	{
-		"inputs": [
-			{
-				"name": "paymentInfo",
-				"type": "tuple",
-				"components": [
-					{"name": "operator", "type": "address"},
-					{"name": "payer", "type": "address"},
-					{"name": "receiver", "type": "address"},
-					{"name": "token", "type": "address"},
-					{"name": "maxAmount", "type": "uint120"},
-					{"name": "preApprovalExpiry", "type": "uint48"},
-					{"name": "authorizationExpiry", "type": "uint48"},
-					{"name": "refundExpiry", "type": "uint48"},
-					{"name": "minFeeBps", "type": "uint16"},
-					{"name": "maxFeeBps", "type": "uint16"},
-					{"name": "feeReceiver", "type": "address"},
-					{"name": "salt", "type": "uint256"}
-				]
-			}
-		],
-		"name": "getHash",
-		"outputs": [{"name": "", "type": "bytes32"}],
-		"stateMutability": "view",
-		"type": "function"
-	}
-]`)
-
-// PaymentInfoTypehash is the EIP-712 typehash for PaymentInfo struct (from contract)
-const PaymentInfoTypehash = "PaymentInfo(address operator,address payer,address receiver,address token,uint120 maxAmount,uint48 preApprovalExpiry,uint48 authorizationExpiry,uint48 refundExpiry,uint16 minFeeBps,uint16 maxFeeBps,address feeReceiver,uint256 salt)"
-
 // Note: MaxUint256() is defined in utils.go
