+++ go/mechanisms/evm/eip712.go
@@ -3,7 +3,6 @@ package evm
 import (
 	"fmt"
 	"math/big"
-	"strings"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/math"
@@ -342,87 +341,6 @@ func HashReceiveWithAuthorization(
 	}
 
 	return HashTypedData(domain, types, "ReceiveWithAuthorization", message)
-}
-
-// ComputeAuthCaptureNonce computes the deterministic payer-agnostic nonce for the authCapture scheme.
-// This mirrors the on-chain _getHashPayerAgnostic logic:
-//
-//	paymentInfoZeroPayer = paymentInfo with payer = address(0)
-//	paymentInfoHash = keccak256(abi.encode(PAYMENT_INFO_TYPEHASH, paymentInfoZeroPayer))
-//	nonce = keccak256(abi.encode(chainId, escrowAddress, paymentInfoHash))
-//
-// Args:
-//
-//	chainID: The chain ID
-//	escrowAddress: The AuthCaptureEscrow contract address
-//	paymentInfo: The AuthCapturePaymentInfo struct
-//
-// Returns:
-//
-//	32-byte nonce as hex string with 0x prefix
-//	error if computation fails
-func ComputeAuthCaptureNonce(chainID *big.Int, escrowAddress string, paymentInfo AuthCapturePaymentInfo) (string, error) {
-	// Compute PaymentInfo typehash
-	typehash := crypto.Keccak256([]byte(PaymentInfoTypehash))
-
-	// Parse PaymentInfo fields
-	operator := common.HexToAddress(paymentInfo.Operator)
-	payer := common.HexToAddress("0x0000000000000000000000000000000000000000") // zeroed payer
-	receiver := common.HexToAddress(paymentInfo.Receiver)
-	token := common.HexToAddress(paymentInfo.Token)
-
-	maxAmount, ok := new(big.Int).SetString(paymentInfo.MaxAmount, 10)
-	if !ok {
-		return "", fmt.Errorf("invalid maxAmount: %s", paymentInfo.MaxAmount)
-	}
-
-	preApprovalExpiry := new(big.Int).SetUint64(paymentInfo.PreApprovalExpiry)
-	authorizationExpiry := new(big.Int).SetUint64(paymentInfo.AuthorizationExpiry)
-	refundExpiry := new(big.Int).SetUint64(paymentInfo.RefundExpiry)
-	minFeeBps := new(big.Int).SetUint64(uint64(paymentInfo.MinFeeBps))
-	maxFeeBps := new(big.Int).SetUint64(uint64(paymentInfo.MaxFeeBps))
-	feeReceiver := common.HexToAddress(paymentInfo.FeeReceiver)
-
-	// Parse salt - can be hex or decimal
-	var salt *big.Int
-	if strings.HasPrefix(paymentInfo.Salt, "0x") || strings.HasPrefix(paymentInfo.Salt, "0X") {
-		salt, ok = new(big.Int).SetString(strings.TrimPrefix(strings.TrimPrefix(paymentInfo.Salt, "0x"), "0X"), 16)
-	} else {
-		salt, ok = new(big.Int).SetString(paymentInfo.Salt, 10)
-	}
-	if !ok {
-		return "", fmt.Errorf("invalid salt: %s", paymentInfo.Salt)
-	}
-
-	// abi.encode(TYPEHASH, operator, payer(zeroed), receiver, token, maxAmount,
-	//            preApprovalExpiry, authorizationExpiry, refundExpiry, minFeeBps, maxFeeBps, feeReceiver, salt)
-	// Each field is padded to 32 bytes
-	encoded := make([]byte, 0, 13*32) // typehash + 12 fields
-	encoded = append(encoded, common.LeftPadBytes(typehash, 32)...)
-	encoded = append(encoded, common.LeftPadBytes(operator.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(payer.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(receiver.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(token.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(maxAmount.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(preApprovalExpiry.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(authorizationExpiry.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(refundExpiry.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(minFeeBps.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(maxFeeBps.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(feeReceiver.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(salt.Bytes(), 32)...)
-
-	paymentInfoHash := crypto.Keccak256(encoded)
-
-	// nonce = keccak256(abi.encode(chainId, escrowAddress, paymentInfoHash))
-	escrow := common.HexToAddress(escrowAddress)
-	nonceEncoded := make([]byte, 0, 3*32)
-	nonceEncoded = append(nonceEncoded, common.LeftPadBytes(chainID.Bytes(), 32)...)
-	nonceEncoded = append(nonceEncoded, common.LeftPadBytes(escrow.Bytes(), 32)...)
-	nonceEncoded = append(nonceEncoded, common.LeftPadBytes(paymentInfoHash, 32)...)
-
-	nonce := crypto.Keccak256(nonceEncoded)
-	return BytesToHex(nonce), nil
 }
 
 // HashPermit2Authorization hashes a PermitWitnessTransferFrom message for Permit2.
